User login is finished and working
There is an endpoint for creating a user, but it has an odd error right
now (it'll throw a 500 error the first time you send the request, but
will respond correctly the second time you send). But, since the point
of this exercise was to get login working, I'm pushing.

The next task will be to fix user creation.